backend/pyproject.toml
[project]
name = "triageiq"
version = "1.0.0"
description = "Production-grade AI support ticket triage system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Validation & settings
    "pydantic>=2.7.0",
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",
    # AI
    "openai>=1.35.0",
    # Auth & security
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    # Redis
    "redis>=5.0.0",
    # Rate limiting
    "slowapi>=0.1.9",
    # Metrics
    "prometheus-fastapi-instrumentator>=7.0.0",
    # Input sanitization
    "bleach>=6.1.0",
    # Logging
    "structlog>=24.2.0",
    "aiosmtplib>=3.0",
    "jinja2>=3.1",
    "firebase-admin>=7.3.0",
    "black>=26.3.1",
    "psycopg[binary,pool]>=3.3.4",
    "python-multipart>=0.0.28",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "ruff>=0.5.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["app", "tests"]
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = [
    "B008",
    "B904",
    "UP042",
    "F821",
]


[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "N"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
warn_return_any = false
exclude = ["alembic/", "tests/"]
mypy_path = "."
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--tb=short", "-q"]

[tool.coverage.run]
source = ["app"]
omit = ["alembic/*", "tests/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.9",
    "types-bleach>=6.3.0.20260402",
]
